Add Follow capabilities (#102)
This adds following capabilities to individual profiles and fixes
various context related view around org vs individual profiles. There is
some refactoring necessary in the profiles to clean up the structure a
bit but this will come in a separate PR as it will be easier to treat it
all at once.