[Design System] `NumberField` supports prefix text (#110)
Adds `prefixText` to `NumberField`. This is a common pattern when
needing a prefix-able text that isn't directly tied to the controlled
value. A use case is to prefix a dollar sign for currency.

The `prefixText` width calculation is dynamic and performant (resize
observer).

---

## Design

<img width="633" height="240" alt="Screenshot 2025-08-23 at 9 06 00 AM"
src="https://github.com/user-attachments/assets/3c6dd30d-9138-4680-9cd8-16180e4a5463"
/>

## Dynamic Width


https://github.com/user-attachments/assets/92c1d670-de5b-4030-b90d-32ad461fe433