Switch carousel for a list
I've started to explore custom carousels. But for the organizations list
on the landing page, I think a simply list would be best for
accessibility and ease of maintenance. A carousel is not needed, we can
get the desired UX on mobile with CSS scroll snapping.

This PR is for that change.