Decision Results API Endpoint (#227)
Adds in an API endpoint for retrieving the results of a decision
process.

A follow-up PR contains the frontend usage of this (#226 )

When we process a result we store the results into a separate table
which becomes the source. We query the results separately in the
endpoint and enrich proposals with some data for each (vote count).