Decision Results Page (#226)
A stacked PR on top of the API endpoint that adds in the page for
displaying a decision process' results. It is missing the stats at the
top but I will do this in a separate PR so we can ship the initial
results page as an MVP.

It also refactors the components a bit.

<img width="1237" height="786" alt="Screenshot 2025-10-29 at 19 52 57"
src="https://github.com/user-attachments/assets/7b41c107-6787-4e72-97f4-8800c7a3cfd3"
/>

There are still visual tweaks needed but this sets up the baseline.

---------

Co-authored-by: Nour Malaeb <nourmalaeb@gmail.com>