Allow viewing All Proposals in voting with filter (#252)
This allows a viewer to browse non-shortlisted proposals while in the
voting phase. They can click to view any updates or comments.

It follows the style of the voting card but doesn't enable the voting
selection on them. We can further iterate if we need to differentiate
them even more.

We also add in backend logic to enforce that only approved proposals are
allowed to be submitted (this will be configurable with the upcoming
pipeline/reducer system).

<img width="1195" height="357" alt="Screenshot 2025-11-04 at 11 45 14"
src="https://github.com/user-attachments/assets/c9d4e245-24e9-442a-82f3-4c19e18bdf74"
/>

Video:


https://github.com/user-attachments/assets/393e192b-31b0-4bec-bb44-838102a59cd2