Update PP Voting Texts (#258)
Updates the hardcoded texts we added for PP so they could have
translations in the first pilot. We don't have translatable
user generated content in the app just yet so this exists as a stopgap
to support the organization as we develop out the ability to include
translatable user generated content.