Updated the FE Decision Permissions to match Backend (#285)
This updates the permissions checks on the frontend to match the
permission check on the backend that looks at the profile owner to
determine whether one can edit a process (usually, admins of an
organization).

This formally would show edit buttons if you have access through one of
your org roles but were in the context of an individual user or another
org role.
It had no affect as the backend checks to make sure that you are in the
context of the owner of the profile (and process).

The editing looked like this:
<img width="1137" height="306" alt="Screenshot 2025-11-13 at 12 52 26"
src="https://github.com/user-attachments/assets/21e9bfd2-ff96-4d77-a01e-f2272e011c12"
/>

And now when switching to your individual profile you will see this like
other individuals would see.
<img width="1116" height="265" alt="Screenshot 2025-11-13 at 12 52 38"
src="https://github.com/user-attachments/assets/1ed3cffa-53da-494e-82e0-04fe00e9f582"
/>