Trim Categories when Entered in the Decisions UI (#299)
A small fix to get a fix in sooner than after we redo parts of this with
some new structures. It updates parts of our prototype code so there are
some type anys in here that can safely be ignored for the moment
(proposal data is untyped).

A newer PR with some refactoring will follow.

Previously we were not trimming categories as they were entered into the
DB causing some issues when filtering. This updates that.