Update Styles on Reaction email (#329)
This updates the styles on the reaction email to match the Figma. It
also moves some colors into the shared tailwind config so that we can
match our @op/ui design system (which causes a circular dependency if we
use it directly AND we need to convert to HEX for usage in react-email).

<img width="602" height="463" alt="Screenshot 2025-11-26 at 12 22 18"
src="https://github.com/user-attachments/assets/fc32c533-deee-4d28-8389-c9a863502502"
/>