Fix SidePanel Disappearing at Breakpoints (#335)
This fixes an issue at small breakpoints where the sidepanel disappears
altogether (and can't be opened).
It also allows us to tie the breakpoint to a known breakpoint in our
configuration.

Previously, this would "open" the panel but the panel would just
disappear. As a fix for now, we open the panel and shift the contents of
the screen to the right since we are not in mobile territory just yet.


https://github.com/user-attachments/assets/8789cda6-2de2-4c75-9072-5b9940c79a9e