Fix SSR Client Fetching With Cookies (#364)
This fixes a Next.js issue with TRPC wherein no cookies are passed to
client queries that are called in a server side render preppas which
calls our TRPC useSuspenseQuery calls.
While it doesn't affect the functionality, it affects performance and
floods our server logs with Auth errors.

It's fixed by encoding cookies that decoded during SSR to avoid auth
issues while still disallowing them to be read in the frontend.