Open Telemetry Logger (#381)
Removes Axiom for a more flexible Open Telemetry logger that we can use
with any OTel provider like SigNoz or Posthog.

Pino was desirable in order to have multiple transports but has some
issues related to edge functions that was complicating it considerably.
So here we use @vercel/otel to deal with that and send traces and logs
to a single OTel provider.

Logs are enriched with traces and a follow-up PR will be made for
WebVitals and client logging.

A followup PR to switch all console.logs to logger.info will also
follow. For now just console.errors are migrated.