Locally handle optimistic update (#392)
We had a regression where optimistic reaction updates on the Landing
Screen PostFeed were not working any longer. We need to optimistically
update this item very fast to feel responsive. Additionally, if we react
to a post on a different screen that appears in the Landing Screen feed,
we need the appropriate reaction to appear immediately.

Server channels will solve this issue, however, until then we are using
a LOCAL optimistic update now (as opposed to a more complex react-query
optimistic update) and needing to trigger updates when data changes from
the backend.


https://github.com/user-attachments/assets/d4a7ba7c-d68a-4508-a176-624e6604c64b